+++ pkg/lib/record_transaction_detail.go
@@ -116,7 +116,7 @@ func (h *TransactionDetail) string(opts ...int64) string {
 
 	for _, composite := range h.Composite {
 		if maxLen > 0 {
-			if buf.Len()+len(composite)+2 > int(maxLen) {
+			if int64(buf.Len()+len(composite)+2) > maxLen {
 				// refresh buf
 				buf.WriteString("/" + "\n") // added new line
 				totalBuf.WriteString(buf.String())
